# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Ecommerce API Documentation",
        "description": "Swagger/OpenAPI documentation for the Laravel ecommerce backend.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local development server"
        }
    ],
    "paths": {
...
        "/api/products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get paginated products",
                "operationId": "listProducts",
                "parameters": [
                    {
                        "name": "category",
                        "in": "query",
                        "description": "Filter by category slug",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "vegetables"
                        }
                    },
                    {
                        "name": "min_price",
                        "in": "query",
                        "description": "Minimum price",
                        "required": false,
                        "schema": {
                            "type": "number",
                            "format": "float",
                            "example": 1
                        }
                    },
                    {
                        "name": "max_price",
                        "in": "query",
                        "description": "Maximum price",
                        "required": false,
                        "schema": {
                            "type": "number",
                            "format": "float",
                            "example": 100
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Sort order",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "latest",
                            "enum": [
                                "latest",
                                "price-asc",
                                "price-desc"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated products",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProductPageResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/products/search": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Search products by keyword",
                "operationId": "searchProducts",
                "parameters": [
                    {
                        "name": "q",
                        "in": "query",
                        "description": "Search keyword",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "chili"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Search results",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProductPageResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/products/{product}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get full product details",
                "operationId": "showProduct",
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "description": "Product ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProductDetailResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/orders": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Get paginated order history",
                "operationId": "listOrders",
                "responses": {
                    "200": {
                        "description": "Paginated order list",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/OrderPageResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Orders"
                ],
                "summary": "Create a new order",
                "operationId": "createOrder",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/OrderCreateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Order created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/OrderCreateResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/orders/{order}": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Get order details",
                "operationId": "showOrder",
                "parameters": [
                    {
                        "name": "order",
                        "in": "path",
                        "description": "Order ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 101
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Order details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/OrderDetailResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ProductPageResponse": {
                "type": "object",
                "example": {
                    "success": true,
                    "data": {
                        "current_page": 1,
                        "data": [
                            {
                                "id": 10,
                                "name": "Red Chili",
                                "slug": "red-chili",
                                "description": "Hot and fresh chili peppers",
                                "price": 2.99,
                                "stock": 50,
                                "image": "products/chili.jpg",
                                "category": {
                                    "id": 1,
                                    "name": "Vegetables",
                                    "slug": "vegetables"
                                }
                            }
                        ],
                        "per_page": 12,
                        "last_page": 1,
                        "total": 1
                    }
                }
            },
            "ProductDetailResponse": {
                "type": "object",
                "example": {
                    "success": true,
                    "data": {
                        "id": 10,
                        "name": "Red Chili",
                        "slug": "red-chili",
                        "description": "Hot and fresh chili peppers",
                        "price": 2.99,
                        "stock": 50,
                        "image": "products/chili.jpg",
                        "category": {
                            "id": 1,
                            "name": "Vegetables",
                            "slug": "vegetables"
                        },
                        "reviews": [
                            {
                                "id": 5,
                                "rating": 5,
                                "comment": "Fresh, tasty and delivered fast.",
                                "created_at": "2026-06-26T10:00:00Z",
                                "user": {
                                    "id": 2,
                                    "name": "Jane Doe",
                                    "email": "jane@example.com"
                                }
                            }
                        ]
                    }
                }
            },
...
            "OrderPageResponse": {
                "type": "object",
                "example": {
                    "success": true,
                    "data": {
                        "current_page": 1,
                        "data": [
                            {
                                "id": 101,
                                "total_price": 25.5,
                                "status": "pending",
                                "shipping_address": "123 Main St, Phnom Penh",
                                "created_at": "2026-06-26T10:00:00Z",
                                "items_count": 3
                            }
                        ],
                        "per_page": 10,
                        "last_page": 1,
                        "total": 1
                    }
                }
            },
            "OrderCreateRequest": {
                "type": "object",
                "example": {
                    "items": [
                        {
                            "product_id": 10,
                            "quantity": 2
                        }
                    ],
                    "shipping_address": "123 Main St, Phnom Penh",
                    "notes": "Please call on arrival"
                }
            },
            "OrderCreateResponse": {
                "type": "object",
                "example": {
                    "success": true,
                    "message": "Order created successfully",
                    "order": {
                        "id": 101,
                        "total_price": 25.5,
                        "status": "pending",
                        "shipping_address": "123 Main St, Phnom Penh",
                        "created_at": "2026-06-26T10:00:00Z",
                        "items": [
                            {
                                "id": 5001,
                                "order_id": 101,
                                "product_id": 10,
                                "quantity": 2,
                                "price": 2.99
                            }
                        ]
                    }
                }
            },
            "OrderDetailResponse": {
                "type": "object",
                "example": {
                    "success": true,
                    "data": {
                        "id": 101,
                        "total_price": 25.5,
                        "status": "pending",
                        "shipping_address": "123 Main St, Phnom Penh",
                        "created_at": "2026-06-26T10:00:00Z",
                        "items": [
                            {
                                "id": 5001,
                                "order_id": 101,
                                "product_id": 10,
                                "quantity": 2,
                                "price": 2.99,
                                "product": {
                                    "id": 10,
                                    "name": "Red Chili",
                                    "slug": "red-chili",
                                    "price": 2.99,
                                    "image": "products/chili.jpg"
                                }
                            }
                        ]
                    }
                }
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Use the token returned from login or register in the Authorization header as Bearer <token>.",
                "bearerFormat": "Token",
                "scheme": "bearer"
            }
        }
    },
...
}